Consolidate packages.yaml code to preferred_packages
